+++ Utilities/DebugConsole/DebugConsole.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using UnityEngine;
@@ -11,8 +12,12 @@ using UnityEngine;
 namespace Sheepy.PhoenixPt.DebugConsole {
 
    public class ModConfig {
+      public int  Config_Version = 20200324;
       public bool Log_Game_Error = true;
       public bool Log_Game_Info = true;
+      public bool Log_Modnix_Error = true;
+      public bool Log_Modnix_Info = true;
+      public bool Log_Modnix_Verbose = false;
    }
 
    public class Mod : ZyMod {
@@ -25,14 +30,57 @@ namespace Sheepy.PhoenixPt.DebugConsole {
          SetApi( api, out Config );
          if ( Config.Log_Game_Error || Config.Log_Game_Info ) {
             Application.logMessageReceived += UnityToConsole;
-            Patch( typeof( TimingScheduler ), "Update", postfix: nameof( BufferToConsole ) );
+            TryPatch( typeof( TimingScheduler ), "Update", postfix: nameof( BufferToConsole ) );
+         }
+         if ( Config.Log_Modnix_Error || Config.Log_Modnix_Info || Config.Log_Modnix_Verbose ) {
+            var loader = api( "assembly", "modnix" ) as Assembly
+               ?? AppDomain.CurrentDomain.GetAssemblies().First( e => e.FullName.StartsWith( "ModnixLoader", StringComparison.Ordinal ) );
+            var type = loader?.GetType( "Sheepy.Logging.FileLogger" );
+            if ( type != null ) {
+               ModnixLogEntryLevel = loader.GetType( "Sheepy.Logging.LogEntry" ).GetField( "Level" );
+               if ( ModnixLogEntryLevel == null ) Warn( "Modnix log level not found. All log forwarded." );
+               TryPatch( type, "ProcessEntry", postfix: nameof( ModnixToConsole ) );
+            } else
+               Info( "Modnix assembly not found, log not forwarded." );
          }
       }
 
-      private static List<string> Buffer = new List<string>();
+      private readonly static List<string> Buffer = new List<string>();
+
+      private static FieldInfo ModnixLogEntryLevel;
+
+      private static void ModnixToConsole ( object entry, string txt ) { try {
+         txt = txt.Trim();
+         if ( txt.Length == 0 || txt.Length > 1000 || txt.Contains( ".ModnixToConsole" ) ) return;
+         if ( ModnixLogEntryLevel != null ) {
+            var level = (TraceEventType) ModnixLogEntryLevel.GetValue( entry );
+            string prefix;
+            lock ( _Lock ) switch ( level ) {
+               case TraceEventType.Critical: case TraceEventType.Error:
+                  if ( ! Config.Log_Modnix_Error ) return;
+                  prefix = "Error ";
+                  break;
+                case TraceEventType.Warning:
+                  if ( ! Config.Log_Modnix_Error ) return;
+                  prefix = "Warning ";
+                  break;
+               case TraceEventType.Information :
+                  if ( ! Config.Log_Modnix_Info ) return;
+                  prefix = "Log [Info] ";
+                  break;
+               default :
+                  if ( ! Config.Log_Modnix_Verbose ) return;
+                  prefix = "Log [Vebo] ";
+                  break;
+            }
+            txt = prefix + txt;
+         }
+         lock ( Buffer ) Buffer.Add( txt );
+      } catch ( Exception ex ) {
+         Error( ex );
+      } }
 
       private static void UnityToConsole ( string condition, string stackTrace, LogType type ) { try {
-         //if ( type == LogType.Error && condition.Contains( "inside a graphic rebuild loop. This is not supported." ) ) return;
          switch ( type ) {
             case LogType.Exception: case LogType.Error: case LogType.Warning:
                if ( ! Config.Log_Game_Error ) return; break;
@@ -40,20 +88,26 @@ namespace Sheepy.PhoenixPt.DebugConsole {
                if ( ! Config.Log_Game_Info ) return; break;
          }
          // Long message can cause the 
-         if ( condition.Length > 1000 || stackTrace?.Contains( "UnityTools.LogFormatter" ) == true || condition.Contains( "Called from a secondary Thread" ) ) return;
+         if ( condition.Length > 1000 || stackTrace.Length > 1000 ||
+              condition.Contains( "Called from a secondary Thread" ) ||
+              stackTrace?.Contains( "UnityTools.LogFormatter" ) == true ) return;
          var line = $"{type} {condition}   {stackTrace}".Trim();
          if ( line.Length == 0 ) return;
-         Buffer.Add( line );
+         lock ( Buffer ) Buffer.Add( line );
       } catch ( Exception ex ) {
          Error( ex );
       } }
 
       private static void BufferToConsole () { try {
-         if ( Buffer.Count == 0 ) return;
+         string[] lines;
+         lock ( Buffer ) {
+            if ( Buffer.Count == 0 ) return;
+            lines = Buffer.ToArray();
+            Buffer.Clear();
+         }
          var console = GameConsoleWindow.Create();
-         foreach ( var line in Buffer )
+         foreach ( var line in lines )
             console.WriteLine( line );
-         Buffer.Clear();
       } catch ( Exception ex ) {
          Error( ex );
       } }